refactor(tui): extract palette runtime component
